[#6] gracefully handle when faces have invalid colors
When we call `color-name-to-rgb` we now catch the case where the color
cannot be properly converted. This occurs when the color is specified
by a name rather than a hex value, but the name isn't known to emacs.
In that situation we ignore that face and don't try to dim it.