work-in-progress: try algorithms for dimming via interpolation
All of these algorithms compute the dimmed version of a face in a way
that is more similar to the "background" in terms of hue, saturation,
and lightness. We're assuming (face-background 'default) is the target
background color. We're using simple linear interpolation as the
transformation.

There are many different colorspaces in which to perform the linear
interpolation but many (all?) of the sophisticated colorspaces also
have the property that it's possible to generate theoretical colors
that are not representable on the target display, be it a LCD, CRT or
projector. This is more likely when interpolating highly saturated
colors at the "limits" of the colorspace, and the result will be
clamped to some other legal color value. In practice, what you'll see
is the "dimmed" color takes on a hue that you would never believe
"sits on the line" between the foreground and background color.

`dimmer-compute-rgb` interpolates in RGB space, this is active
`dimmer-compute-rgb-in-HSL-space` interpolates in HSL
`dimmer-compute-rgb-in-LAB-space` interpolates in CIELAB

Edit the code to experiment with a different algorithms.

Or better, try the interactive visualization of gradient ramps in
various colorspaces

http://davidjohnstone.net/pages/lch-lab-colour-gradient-picker