Calls dimmer-process-all asynchronously...
Change: dimmer-config-change-hook and dimmer-command-hook now both
call dimmer-process-all asynchronously. Additionally, before running
they check for the existence of an unrealized but queued call to
dimmer-process-all. This means that we are able to respond to
operations that alter the window configuration multiple times with a
single call to dimmer-process-all, greatly reducing the amount of work
dimmer has to do.

Change: on activation, dimmer mode makes a single eager call to
dimmer-process-all, to correctly set the initial state. On
deactivation, dimmer-restore-all is now called asynchronously in order
to account for any unrealized calls to dimmer-process-all.