set permanent-local attribute on dimmer-buffer-face-remaps
Fix #33

Users who install org-plus-contrib have reported a bug where the "*Org
Agenda*" buffer is stuck in a dimmed state.

Debugging revealed that `org-agenda` causes the "*Org Agenda*" buffer
to be created, dimmed, and then undimmed with the contents filled in.

When running org-plus-contrib, somewhere in the middle of this process
the buffer local variable `dimmer-buffer-face-remaps` (which is used
to track if a buffer has been dimmed) is reset to nil so that the
buffer isn't seen as "dimmed" even though the buffer has its faces
remapped.

The underlying issue is that the buffer local variable can be forced
to nil under certain circumstances, but we never want that. Setting
the permanent-local attribute on `dimmer-buffer-face-remaps` prevents
its value from being killed, e.g. if there is a major mode change in
the buffer.